colors and gradiants

+++ src/utils/vectors/vect2d/vect2d.c
@@ -6,13 +6,13 @@
 /*   By: olimarti <olimarti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 00:51:11 by olimarti          #+#    #+#             */
-/*   Updated: 2023/03/23 04:25:59 by olimarti         ###   ########.fr       */
+/*   Updated: 2023/03/28 08:31:43 by olimarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../utils.h"
 
-t_vect2d	vect2d(float x, float y)
+t_vect2d	vect2d(double x, double y)
 {
 	t_vect2d	result;
